SceneManagerの内容をgameに移行させました
