+++ GameProject/Game.h
@@ -1,81 +1,47 @@
 ﻿#pragma once
-#include<vector>
+#include"DxLib.h"
+#include<list>
+
 
 //クラスのプロトタイプ宣言
-class Player;
-class Map;
-class Gimmick;
-class Utility;
-class UI;
-class BgModel;
-class BG;
-class Cloud;
-class Player3DDraw;
-class Gimmick3D;
-class GroundDraw3D;
-class Effect;
-class EnemyManager;
-class ShotManager;
-class Camera;
+class SceneBase;
+class FrameRate;
 
-/// <summary>
-/// ゲーム全体のインスタンスなどをまとめたクラス
-/// </summary>
 class Game
 {
 public:
-    // ゲームの状態.
-    enum STATE
-    {
-        STATE_TITLE,		// タイトル
-        STATE_GAME,			// ゲーム中
-        STATE_GAMECLEAR,	//クリア
-        STATE_GAMEOVER,		// ゲームオーバー
-    };
-	//コンストラクタ・デストラクタ
-	Game();
-	~Game();
-	void GameStateChange();		//ゲームステータスを移行する際の処理
-	void Initialize();			//初期化
-	//メイン描画の後にサブの処理をする必要があるので別々にする
-	void Update();				//メイン画面の更新処理
-	void Draw();				//メイン画面の描画
-    /// <summary>
-    /// フレームレートの調整をする変数などの初期化
-    /// </summary>
-    void InitializeFrameRate();
-    /// <summary>
-    /// フレームレートの計算のためにカウントなどを取得
-    /// </summary>
-    void UpdateFrameRate();
-    /// <summary>
-    /// 取得したカウントを使用し実際にフレームレートを調節する
-    /// </summary>
-    void ControlFrameRate();
-private:
-    //static変数
-    //FPSコントロール用
-    int startTime;           //ループ開始時の時間
-    int roopCount;           //ループ回数
+    //コンストラクト
+    Game();
+    //デストラクタ
+    ~Game();
 
-	//プレイヤーやマップなどのクラスをまとめる
-	Player* player;
-	Utility* utility;
-	UI* ui;
-	Camera* camera;
-	Map* map;
-	EnemyManager* enemyManager;
-    ShotManager* shotManager;
-	//ゲームのステータス関連（スクロールスピードなど)
-	
-	bool keyRelease = false;
-	bool keyOn = false;
-	bool prevKeyOn = false;
-	int effectPlayCount = 0;
-	bool keyStop = false;
-    //ゲームの状態
-    int gameState = STATE_GAME;			//ゲームステート（タイトル、ゲーム中など)
+    /// <summary>
+    /// ゲームループをまわす処理
+    /// </summary>
+    void GameLoop();
+
+    /// <summary>
+    /// ゲームを継続するかの判断
+    /// </summary>
+    /// <returns>継続するかのフラグ</returns>
+    bool IsContinueGame();
+
+
+private:
+
+    /// <summary>
+    /// 現在のシーンを削除して次のシーンのポインタを入れる
+    /// </summary>
+    void SceneChange();
+
+    SceneBase* nowScene;        //現在のループでのシーンを保存するポインタ
+    SceneBase* nextScene;       //次のループでのシーンを保存するポインタ
+
+    FrameRate* frameRate;       //フレームレート計算用クラス
+
+
 
 
 
 };
+
